use ints instead of floats

+++ w1_sobel/src/main.rs
@@ -1,8 +1,8 @@
 use image::{io::Reader as ImageReader, ImageBuffer, Luma};
 use std::{env, time::Instant};
 
-const SOBEL_X: [[f64; 3]; 3] = [[-1.0, -2.0, -1.0], [0.0, 0.0, 0.0], [1.0, 2.0, 1.0]];
-const SOBEL_Y: [[f64; 3]; 3] = [[-1.0, 0.0, 1.0], [-2.0, 0.0, 2.0], [-1.0, 0.0, 1.0]];
+const SOBEL_X: [[i16; 3]; 3] = [[-1, -2, -1], [0, 0, 0], [1, 2, 1]];
+const SOBEL_Y: [[i16; 3]; 3] = [[-1, 0, 1], [-2, 0, 2], [-1, 0, 1]];
 
 //  cargo run -- img/house.jpg img/ball.jpg img/lizard.jpg
 fn main() {
@@ -39,22 +39,22 @@ fn sobel(img: &ImageBuffer<Luma<u8>, Vec<u8>>) -> ImageBuffer<Luma<u8>, Vec<u8>>
 
 fn conv_3x3(
     img: &ImageBuffer<Luma<u8>, Vec<u8>>,
-    mask: [[f64; 3]; 3],
+    mask: [[i16; 3]; 3],
 ) -> ImageBuffer<Luma<u8>, Vec<u8>> {
     let (w, h) = img.dimensions();
     let mut output = ImageBuffer::new(w, h);
 
     for x in 1..(w - 1) {
         for y in 1..(h - 1) {
-            let p1 = img.get_pixel(x - 1, y - 1).0[0] as f64 * mask[0][0];
-            let p2 = img.get_pixel(x - 1, y).0[0] as f64 * mask[0][1];
-            let p3 = img.get_pixel(x - 1, y + 1).0[0] as f64 * mask[0][2];
-            let p4 = img.get_pixel(x, y - 1).0[0] as f64 * mask[1][0];
-            let p5 = img.get_pixel(x, y).0[0] as f64 * mask[1][1];
-            let p6 = img.get_pixel(x, y + 1).0[0] as f64 * mask[1][2];
-            let p7 = img.get_pixel(x + 1, y - 1).0[0] as f64 * mask[2][0];
-            let p8 = img.get_pixel(x + 1, y).0[0] as f64 * mask[2][1];
-            let p9 = img.get_pixel(x + 1, y + 1).0[0] as f64 * mask[2][2];
+            let p1 = img.get_pixel(x - 1, y - 1).0[0] as i16 * mask[0][0];
+            let p2 = img.get_pixel(x - 1, y).0[0] as i16 * mask[0][1];
+            let p3 = img.get_pixel(x - 1, y + 1).0[0] as i16 * mask[0][2];
+            let p4 = img.get_pixel(x, y - 1).0[0] as i16 * mask[1][0];
+            let p5 = img.get_pixel(x, y).0[0] as i16 * mask[1][1];
+            let p6 = img.get_pixel(x, y + 1).0[0] as i16 * mask[1][2];
+            let p7 = img.get_pixel(x + 1, y - 1).0[0] as i16 * mask[2][0];
+            let p8 = img.get_pixel(x + 1, y).0[0] as i16 * mask[2][1];
+            let p9 = img.get_pixel(x + 1, y + 1).0[0] as i16 * mask[2][2];
             let val: u8 = (p1 + p2 + p3 + p4 + p5 + p6 + p7 + p8 + p9).abs() as u8;
             output.put_pixel(x, y, Luma([val]));
         }
